test(web): centralize isolated web runner
